Changed the signature of `wread' it should have a `size_t' and not a int to be conform to the C ANSI specification
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@9919 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/eif_update.h
@@ -52,7 +52,7 @@ extern short wshort(void);
 extern long wlong(void);
 extern int32 wint32(void);
 extern uint32 wuint32(void);
-extern void wread(char *buffer, int nbytes);
+extern void wread(char *buffer, size_t nbytes);
 
 #ifdef __cplusplus
 }
